110331 files, pagetables

+++ fs/inode.h
@@ -11,7 +11,7 @@
 
 struct inode *iget (struct super *sb, size_t inum);
 void iput (struct inode *inode);
-size_t bmap (struct inode *inode, size_t offset);
+size_t bmap (struct inode *inode, off_t offset);
